backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
groq
scikit-learn
pandas
numpy
joblib
pyarrow
requests
python-dotenv
scipy